+++ src/turbo_server.cpp
@@ -1,3 +1,5 @@
+#include "turbo_server.h"
+
 extern "C" {
   #include <unistd.h>
   #include <wayland-server-core.h>
@@ -15,6 +17,7 @@ extern "C" {
   #include <wlr/types/wlr_seat.h>
   #include <wlr/types/wlr_xcursor_manager.h>
   #include <wlr/types/wlr_xdg_shell.h>
+  #include <wlr/xwayland.h>
   #include <wlr/util/log.h>
   #include <wlr/backend/libinput.h>
   #include <xkbcommon/xkbcommon.h>
@@ -22,14 +25,447 @@ extern "C" {
 
 #include <iostream>
 
-#include "turbo_server.h"
 #include "turbo_keyboard.h"
 #include "turbo_view.h"
-#include "turbo_cursor_mode.h"
+#include "turbo_view_xwayland.h"
+#include "turbo_view_xdg.h"
+#include "turbo_output.h"
 
 turbo_server::turbo_server()
   : cursor_mode(TURBO_CURSOR_NONE) {
 
+}
+
+static void new_input_notify(wl_listener *listener, void *data) {
+  /* This event is raised by the backend when a new input device becomes
+   * available. */
+  turbo_server *server = wl_container_of(listener, server, new_input);
+  auto device = static_cast<struct wlr_input_device *>(data);
+
+  switch (device->type) {
+  case WLR_INPUT_DEVICE_KEYBOARD:
+    server->new_keyboard(device);
+    break;
+
+  case WLR_INPUT_DEVICE_POINTER:
+    server->new_pointer(device);
+    break;
+
+  default:
+    break;
+
+  }
+  /* We need to let the wlr_seat know what our capabilities are, which is
+   * communiciated to the client. In TinyWL we always have a cursor, even if
+   * there are no pointer devices, so we always include that capability. */
+  uint32_t caps = WL_SEAT_CAPABILITY_POINTER;
+  if (!wl_list_empty(&server->keyboards)) {
+    caps |= WL_SEAT_CAPABILITY_KEYBOARD;
+  }
+
+  wlr_seat_set_capabilities(server->seat, caps);
+}
+
+static void seat_request_cursor_notify(wl_listener *listener, void *data) {
+  turbo_server *server = wl_container_of(listener, server, request_cursor);
+  /* This event is rasied by the seat when a client provides a cursor image */
+  auto event = static_cast<wlr_seat_pointer_request_set_cursor_event*>(data);
+  wlr_seat_client *focused_client = server->seat->pointer_state.focused_client;
+  /* This can be sent by any client, so we check to make sure this one is
+   * actually has pointer focus first. */
+  if (focused_client == event->seat_client) {
+    /* Once we've vetted the client, we can tell the cursor to use the
+     * provided surface as the cursor image. It will set the hardware cursor
+     * on the output that it's currently on and continue to do so as the
+     * cursor moves between outputs. */
+    wlr_cursor_set_surface(server->cursor, event->surface, event->hotspot_x, event->hotspot_y);
+  }
+}
+
+static void cursor_motion_notify(wl_listener *listener, void *data) {
+  /* This event is forwarded by the cursor when a pointer emits a _relative_
+   * pointer motion event (i.e. a delta) */
+  turbo_server *server = wl_container_of(listener, server, cursor_motion);
+  auto event = static_cast<struct wlr_event_pointer_motion*>(data);
+  /* The cursor doesn't move unless we tell it to. The cursor automatically
+   * handles constraining the motion to the output layout, as well as any
+   * special configuration applied for the specific input device which
+   * generated the event. You can pass NULL for the device if you want to move
+   * the cursor around without any input. */
+  wlr_cursor_move(server->cursor, event->device, event->delta_x, event->delta_y);
+  server->process_cursor_motion(event->time_msec);
+}
+
+static void cursor_motion_absolute_notify(wl_listener *listener, void *data) {
+  /* This event is forwarded by the cursor when a pointer emits an _absolute_
+   * motion event, from 0..1 on each axis. This happens, for example, when
+   * wlroots is running under a Wayland window rather than KMS+DRM, and you
+   * move the mouse over the window. You could enter the window from any edge,
+   * so we have to warp the mouse there. There is also some hardware which
+   * emits these events. */
+  turbo_server *server = wl_container_of(listener, server, cursor_motion_absolute);
+  auto event = static_cast<struct wlr_event_pointer_motion_absolute*>(data);
+  wlr_cursor_warp_absolute(server->cursor, event->device, event->x, event->y);
+  server->process_cursor_motion(event->time_msec);
+}
+
+static void cursor_button_notify(wl_listener *listener, void *data) {
+  /* This event is forwarded by the cursor when a pointer emits a button
+   * event. */
+  turbo_server *server = wl_container_of(listener, server, cursor_button);
+  auto event = static_cast<struct wlr_event_pointer_button*>(data);
+  /* Notify the client with pointer focus that a button press has occurred */
+  wlr_seat_pointer_notify_button(server->seat, event->time_msec, event->button, event->state);
+
+  double sx, sy;
+  wlr_surface *surface;
+  turbo_view *view = server->desktop_view_at(server->cursor->x, server->cursor->y, &surface, &sx, &sy);
+
+  if (event->state == WLR_BUTTON_RELEASED) {
+    /* If you released any buttons, we exit interactive move/resize mode. */
+    server->cursor_mode = TURBO_CURSOR_PASSTHROUGH;
+  } else {
+    if (view != NULL) {
+      /* Focus that client if the button was _pressed_ */
+      view->focus_view(surface);
+    }
+  }
+}
+
+static void cursor_axis_notify(wl_listener *listener, void *data) {
+  /* This event is forwarded by the cursor when a pointer emits an axis event,
+   * for example when you move the scroll wheel. */
+  turbo_server *server = wl_container_of(listener, server, cursor_axis);
+  auto event = static_cast<wlr_event_pointer_axis*>(data);
+  /* Notify the client with pointer focus of the axis event. */
+  wlr_seat_pointer_notify_axis(server->seat, event->time_msec, event->orientation,
+    event->delta, event->delta_discrete, event->source);
+}
+
+static void cursor_frame_notify(wl_listener *listener, void *data) {
+  /* This event is forwarded by the cursor when a pointer emits an frame
+   * event. Frame events are sent after regular pointer events to group
+   * multiple events together. For instance, two axis events may happen at the
+   * same time, in which case a frame event won't be sent in between. */
+  turbo_server *server = wl_container_of(listener, server, cursor_frame);
+  /* Notify the client with pointer focus of the frame event. */
+  wlr_seat_pointer_notify_frame(server->seat);
+}
+
+static void output_frame_notify(wl_listener *listener, void *data) {
+  /* This function is called every time an output is ready to display a frame,
+   * generally at the output's refresh rate (e.g. 60Hz). */
+  turbo_output *output = wl_container_of(listener, output, frame);
+  output->render();
+}
+
+static void new_output_notify(wl_listener *listener, void *data) {
+  /* This event is rasied by the backend when a new output (aka a display or
+   * monitor) becomes available. */
+  turbo_server *server = wl_container_of(listener, server, new_output);
+  auto wlr_output = static_cast<struct wlr_output*>(data);
+
+  wlr_log(WLR_ERROR, "%s connected", wlr_output->name);
+
+  /* Some backends don't have modes. DRM+KMS does, and we need to set a mode
+   * before we can use the output. The mode is a tuple of (width, height,
+   * refresh rate), and each monitor supports only a specific set of modes. We
+   * just pick the monitor's preferred mode, a more sophisticated compositor
+   * would let the user configure it. */
+  if (!wl_list_empty(&wlr_output->modes)) {
+    struct wlr_output_mode *mode = wlr_output_preferred_mode(wlr_output);
+    wlr_output_set_mode(wlr_output, mode);
+    wlr_output_enable(wlr_output, true);
+    if (!wlr_output_commit(wlr_output)) {
+      wlr_log(WLR_ERROR, "Failed to commit output");
+      return;
+    }
+  }
+
+  if (strcmp(wlr_output->name, "X11-1") == 0) {
+    wlr_output_set_scale(wlr_output, 2);
+  }
+
+  if (strcmp(wlr_output->name, "eDP-1") == 0) {
+    wlr_output_set_scale(wlr_output, 3);
+  }
+
+  if (strcmp(wlr_output->name, "DP-0") == 0
+   || strcmp(wlr_output->name, "DP-1") == 0
+   || strcmp(wlr_output->name, "DP-2") == 0) {
+    wlr_output_set_scale(wlr_output, 1);
+  }
+
+  /* Allocates and configures our state for this output */
+  turbo_output *output = new turbo_output();
+  output->wlr_output = wlr_output;
+  output->server = server;
+
+  /* Sets up a listener for the frame notify event. */
+  output->frame.notify = output_frame_notify;
+  wl_signal_add(&wlr_output->events.frame, &output->frame);
+
+  /* Adds this to the output layout. The add_auto function arranges outputs
+   * from left-to-right in the order they appear. A more sophisticated
+   * compositor would let the user configure the arrangement of outputs in the
+   * layout. */
+  wlr_output_layout_add_auto(server->output_layout, wlr_output);
+
+  /* Creating the global adds a wl_output global to the display, which Wayland
+   * clients can see to find out information about the output (such as
+   * DPI, scale factor, manufacturer, etc). */
+  wlr_output_create_global(wlr_output);
+
+  wl_list_insert(&server->outputs, &output->link);
+}
+
+static void xdg_surface_map_notify(wl_listener *listener, void *data) {
+  /* Called when the surface is mapped, or ready to display on-screen. */
+  turbo_view *view = wl_container_of(listener, view, map);
+  turbo_server *server = view->server;
+  view->mapped = true;
+
+  server->position_view(view);
+  view->focus();
+}
+
+static void xdg_surface_unmap_notify(wl_listener *listener, void *data) {
+  /* Called when the surface is unmapped, and should no longer be shown. */
+  turbo_view *view = wl_container_of(listener, view, unmap);
+  view->mapped = false;
+  turbo_server *server = view->server;
+  server->pop_view(view);
+}
+
+static void xdg_surface_destroy_notify(wl_listener *listener, void *data) {
+  /* Called when the surface is destroyed and should never be shown again. */
+  turbo_view *view = wl_container_of(listener, view, destroy);
+  wl_list_remove(&view->link);
+  delete view;
+}
+
+static void xdg_toplevel_request_move_notify(wl_listener *listener, void *data) {
+  /* This event is raised when a client would like to begin an interactive
+   * move, typically because the user clicked on their client-side
+   * decorations. Note that a more sophisticated compositor should check the
+   * provied serial against a list of button press serials sent to this
+   * client, to prevent the client from requesting this whenever they want. */
+  turbo_view *view = wl_container_of(listener, view, request_move);
+  view->begin_interactive(TURBO_CURSOR_MOVE, 0);
+}
+
+static void xdg_toplevel_request_resize_notify(wl_listener *listener, void *data) {
+  /* This event is raised when a client would like to begin an interactive
+   * resize, typically because the user clicked on their client-side
+   * decorations. Note that a more sophisticated compositor should check the
+   * provied serial against a list of button press serials sent to this
+   * client, to prevent the client from requesting this whenever they want. */
+  auto event = static_cast<wlr_xdg_toplevel_resize_event*>(data);
+  turbo_view *view = wl_container_of(listener, view, request_resize);
+  view->begin_interactive(TURBO_CURSOR_RESIZE, event->edges);
+}
+
+static void xdg_toplevel_request_maximize_notify(wl_listener *listener, void *data) {
+  turbo_view *view = wl_container_of(listener, view, request_maximize);
+  view->toggle_maximize();
+}
+
+static void xwayland_surface_request_configure_notify(wl_listener *listener, void *data) {
+  turbo_view_xwayland *view = wl_container_of(listener, view, request_configure);
+  auto event = static_cast<wlr_xwayland_surface_configure_event*>(data);
+  view->x = event->x;
+  view->y = event->y;
+  wlr_xwayland_surface_configure(event->surface, event->x, event->y, event->width, event->height);
+}
+
+static void new_xwayland_surface_notify(wl_listener *listener, void *data) {
+  turbo_server *server = wl_container_of(listener, server, new_xwayland_surface);
+  auto xwayland_surface = static_cast<wlr_xwayland_surface*>(data);
+
+  turbo_view_xwayland *view = new turbo_view_xwayland();
+
+  view->x = 100;
+  view->y = 100;
+  view->server = server;
+  view->xwayland_surface = xwayland_surface;
+  // view->surface_type = TURBO_XWAYLAND_SURFACE;
+
+  view->map.notify = xdg_surface_map_notify;
+  wl_signal_add(&xwayland_surface->events.map, &view->map);
+
+  view->unmap.notify = xdg_surface_unmap_notify;
+  wl_signal_add(&xwayland_surface->events.unmap, &view->unmap);
+
+  view->destroy.notify = xdg_surface_destroy_notify;
+  wl_signal_add(&xwayland_surface->events.destroy, &view->destroy);
+
+  view->request_configure.notify = xwayland_surface_request_configure_notify;
+  wl_signal_add(&xwayland_surface->events.request_configure, &view->request_configure);
+
+  view->request_move.notify = xdg_toplevel_request_move_notify;
+  wl_signal_add(&xwayland_surface->events.request_move, &view->request_move);
+
+  view->request_resize.notify = xdg_toplevel_request_resize_notify;
+  wl_signal_add(&xwayland_surface->events.request_resize, &view->request_resize);
+
+  view->request_maximize.notify = xdg_toplevel_request_maximize_notify;
+  wl_signal_add(&xwayland_surface->events.request_maximize, &view->request_maximize);
+
+   /* Add it to the list of views. */
+  wl_list_insert(&server->views, &view->link);
+}
+
+static void new_xdg_surface_notify(wl_listener *listener, void *data) {
+  /* This event is raised when wlr_xdg_shell receives a new xdg surface from a
+   * client, either a toplevel (application window) or popup. */
+  turbo_server *server = wl_container_of(listener, server, new_xdg_surface);
+  auto xdg_surface = static_cast<wlr_xdg_surface*>(data);
+  if (xdg_surface->role != WLR_XDG_SURFACE_ROLE_TOPLEVEL) {
+    return;
+  }
+
+  /* Allocate a turbo_view for this surface */
+  turbo_view_xdg *view = new turbo_view_xdg();
+  view->x = 0;
+  view->y = 0;
+  view->server = server;
+  view->xdg_surface = xdg_surface;
+  // view->surface_type = TURBO_XDG_SURFACE;
+
+  /* Listen to the various events it can emit */
+  view->map.notify = xdg_surface_map_notify;
+  wl_signal_add(&xdg_surface->events.map, &view->map);
+
+  view->unmap.notify = xdg_surface_unmap_notify;
+  wl_signal_add(&xdg_surface->events.unmap, &view->unmap);
+
+  view->destroy.notify = xdg_surface_destroy_notify;
+  wl_signal_add(&xdg_surface->events.destroy, &view->destroy);
+
+  /* cotd */
+  struct wlr_xdg_toplevel *toplevel = xdg_surface->toplevel;
+
+  view->request_move.notify = xdg_toplevel_request_move_notify;
+  wl_signal_add(&toplevel->events.request_move, &view->request_move);
+
+  view->request_resize.notify = xdg_toplevel_request_resize_notify;
+  wl_signal_add(&toplevel->events.request_resize, &view->request_resize);
+
+  view->request_maximize.notify = xdg_toplevel_request_maximize_notify;
+  wl_signal_add(&toplevel->events.request_maximize, &view->request_maximize);
+
+  /* Add it to the list of views. */
+  wl_list_insert(&server->views, &view->link);
+}
+
+void turbo_server::run() {
+  wlr_log_init(WLR_DEBUG, NULL);
+
+  wl_display = wl_display_create();
+
+  backend = wlr_backend_autocreate(wl_display, NULL);
+
+  renderer = wlr_backend_get_renderer(backend);
+  wlr_renderer_init_wl_display(renderer, wl_display);
+
+
+  wlr_compositor *compositor = wlr_compositor_create(wl_display, renderer);
+  wlr_data_device_manager_create(wl_display);
+
+  output_layout = wlr_output_layout_create();
+
+  wl_list_init(&outputs);
+  new_output.notify = new_output_notify;
+  wl_signal_add(&backend->events.new_output, &new_output);
+
+  wl_list_init(&views);
+  xdg_shell = wlr_xdg_shell_create(wl_display);
+  new_xdg_surface.notify = new_xdg_surface_notify;
+  wl_signal_add(&xdg_shell->events.new_surface, &new_xdg_surface);
+
+  cursor = wlr_cursor_create();
+  wlr_cursor_attach_output_layout(cursor, output_layout);
+
+  cursor_mgr = wlr_xcursor_manager_create(NULL, 24);
+  wlr_xcursor_manager_load(cursor_mgr, 1);
+  wlr_xcursor_manager_load(cursor_mgr, 2);
+  wlr_xcursor_manager_load(cursor_mgr, 3);
+
+  cursor_motion.notify = cursor_motion_notify;
+  wl_signal_add(&cursor->events.motion, &cursor_motion);
+
+  cursor_motion_absolute.notify = cursor_motion_absolute_notify;
+  wl_signal_add(&cursor->events.motion_absolute, &cursor_motion_absolute);
+
+  cursor_button.notify = cursor_button_notify;
+  wl_signal_add(&cursor->events.button, &cursor_button);
+
+  cursor_axis.notify = cursor_axis_notify;
+  wl_signal_add(&cursor->events.axis, &cursor_axis);
+
+  cursor_frame.notify = cursor_frame_notify;
+  wl_signal_add(&cursor->events.frame, &cursor_frame);
+
+  wl_list_init(&keyboards);
+
+  new_input.notify = new_input_notify;
+  wl_signal_add(&backend->events.new_input, &new_input);
+
+  seat = wlr_seat_create(wl_display, "seat0");
+
+  request_cursor.notify = seat_request_cursor_notify;
+  wl_signal_add(&seat->events.request_set_cursor, &request_cursor);
+
+  const char *socket = wl_display_add_socket_auto(wl_display);
+  if (!socket) {
+    wlr_backend_destroy(backend);
+    return;
+  }
+
+  if (!wlr_backend_start(backend)) {
+    wlr_backend_destroy(backend);
+    wl_display_destroy(wl_display);
+    return;
+  }
+
+  setenv("WAYLAND_DISPLAY", socket, true);
+
+  wlr_data_device_manager_create(wl_display);
+
+  xwayland = wlr_xwayland_create(wl_display, compositor, true);
+  wlr_xwayland_set_seat(xwayland, seat);
+
+  new_xwayland_surface.notify = new_xwayland_surface_notify;
+  wl_signal_add(&xwayland->events.new_surface, &new_xwayland_surface);
+  wlr_log(WLR_INFO, "XWayland DISPLAY=%s", xwayland->display_name);
+
+  wlr_log(WLR_INFO, "Wayland WAYLAND_DISPLAY=%s", socket);
+  wl_display_run(wl_display);
+}
+
+void turbo_server::destroy() {
+  wlr_log(WLR_ERROR, "Quitting!");
+
+  wlr_xcursor_manager_destroy(cursor_mgr);
+
+  // wlr_xwayland_destroy(xwayland);
+  // wlr_seat_destroy(seat);
+  // wlr_xcursor_manager_destroy(cursor_mgr);
+  wlr_cursor_destroy(cursor);
+  wlr_output_layout_destroy(output_layout);
+
+  turbo_keyboard *keyboard, *tmp;
+  wl_list_for_each_safe(keyboard, tmp, &keyboards, link) {
+    delete keyboard;
+  }
+
+  /* Once wl_display_run returns, we shut down the  */
+  wl_display_destroy_clients(wl_display);
+
+  // wlr_renderer_destroy(renderer);
+  wlr_backend_destroy(backend);
+
+  wl_display_destroy(wl_display);
 }
 
 static bool handle_alt_keybinding(turbo_server *server, xkb_keysym_t sym) {
